Restructure folders again.
