+++ .gitignore
@@ -2,3 +2,4 @@ temp
 .venv
 temp_
 rocket_league_replay_parser.egg-info
+*.pyc
