+++ replay_parser/__init__.py
@@ -1,0 +1,3 @@
+from .replay_parser import *
+
+VERSION = (0, 1, 0)
